set up delete icon
